Renamed variables and code format
