+++ src/main/java/com/hubspot/jinjava/interpret/UnknownTokenException.java
@@ -2,17 +2,17 @@ package com.hubspot.jinjava.interpret;
 
 public class UnknownTokenException extends InterpretException {
 
-    private final String token;
+  private final String token;
 
-    public UnknownTokenException(String token, int lineNumber) {
-        super("Unknown token found: " + token, lineNumber);
-        this.token = token;
-    }
+  public UnknownTokenException(String token, int lineNumber) {
+    super("Unknown token found: " + token, lineNumber);
+    this.token = token;
+  }
 
-    public String getToken() {
-        return token;
-    }
+  public String getToken() {
+    return token;
+  }
 
-    private static final long serialVersionUID = -388757722051666198L;
+  private static final long serialVersionUID = -388757722051666198L;
 
 }
